Change to use bow.state.Response.Request.URL in Url()
Prefer to use bow.state.Response.URL, because in case of redirects,
the current context depends on the final Request.

A concrete example would be something like OAuth flow, where you
access host A which delegates to host B to do the authorization
via a redirect. If you don't use the Response URL, then the next
Form object may resolve its URL in the context of host A, not B